Point 8. Help Neville find a room

+++ src/main/java/com/codecool/hogwartshouses/model/Student.java
@@ -29,7 +29,7 @@ public class Student {
     private HouseType houseType;
 
     @Column(name = "room_id_for_student")
-    private int roomIdForStudent;
+    private Long roomIdForStudent;
 
 
 }
